refactor(worker): share pagehide ownership
