+++ packages/cli/src/serve/worker/index.ts
@@ -43,6 +43,7 @@ export {
   disconnectClient,
   type DisconnectClientOptions,
 } from './client/disconnect.js';
+export { ownClientUntilPagehide } from './client/pagehide.js';
 
 export {
   // Bridge-peer relay seams (leaf-safe: ids are re-minted in-page, frames go
